refactor(oh7): Lot 2a — move SqliteLog out of storage into bin es_log module (#252)
Relocates `SqliteLog` (implements `core::orchestration::es::log::EventLog`)
from `src/storage/es_log.rs` to `src/es_log.rs`, a bin-root module, breaking
the `storage -> core::es` edge so `storage` can later become a leaf crate.

Pure relocation, no logic change:
- `git mv src/storage/es_log.rs src/es_log.rs`
- `src/storage/mod.rs`: drop `pub mod es_log;`
- `src/main.rs`: add `#[cfg(feature = "storage")] mod es_log;`
- Consumers (`src/cli/{run,run_es_record,run_replay,projections}.rs`):
  `crate::storage::es_log::SqliteLog` -> `crate::es_log::SqliteLog`
