+++ src/main.rs
@@ -1,6 +1,8 @@
 mod audit;
 mod cli;
 mod core;
+#[cfg(feature = "storage")]
+mod es_log;
 mod linker;
 mod logging;
 mod model_registry;
