+++ src/storage/mod.rs
@@ -1,5 +1,3 @@
-#[cfg(feature = "storage")]
-pub mod es_log;
 pub mod queries;
 pub mod schema;
